feat(react-native): floating and a snapping animated local video view
