+++ packages/react-native-sdk/package.json
@@ -39,7 +39,9 @@
     "react-native": ">=0.67.0",
     "react-native-callkeep": ">=4.3.11",
     "react-native-device-info": ">=10.6.0",
+    "react-native-gesture-handler": ">=2.8.0",
     "react-native-incall-manager": ">=4.1.0",
+    "react-native-reanimated": ">=2.7.0",
     "react-native-svg": ">=13.6.0",
     "react-native-voip-push-notification": ">=3.3.1",
     "react-native-webrtc": ">=111.0.3"
@@ -55,6 +57,12 @@
       "optional": true
     },
     "react-native-voip-push-notification": {
+      "optional": true
+    },
+    "react-native-gesture-handler": {
+      "optional": true
+    },
+    "react-native-reanimated": {
       "optional": true
     }
   },
@@ -77,7 +85,9 @@
     "react-native": "0.71.8",
     "react-native-callkeep": "4.3.11",
     "react-native-device-info": "^10.6.0",
+    "react-native-gesture-handler": "2.8.0",
     "react-native-incall-manager": "^4.0.0",
+    "react-native-reanimated": "2.7.0",
     "react-native-svg": "^13.6.0",
     "react-native-voip-push-notification": "3.3.1",
     "react-native-webrtc": "~106.0.4",
