+++ packages/react-native-sdk/src/components/participants/ParticipantView.tsx
@@ -247,6 +247,7 @@ export const ParticipantView = (props: ParticipantViewProps) => {
       </View>
       {canShowVideo ? (
         <VideoRenderer
+          // zOrder should lower than the zOrder used in the floating LocalParticipantView
           zOrder={Z_INDEX.IN_BACK}
           mirror={mirror}
           mediaStream={videoStream}
